refactor assets into different files
